+++ Cargo.toml
@@ -20,13 +20,13 @@ license = "Apache-2.0"
 readme = "README.md"
 repository = "https://github.com/FL03/variants.git"
 rust-version = "1.85.0"
-version = "0.0.1"
+version = "0.0.2"
 
 [workspace.dependencies]
 # local
-variants = { path = "variants", version = "0.0.1" }
-variants-derive = { path = "derive", version = "0.0.1" }
-variants-macros = { path = "macros", version = "0.0.1" }
+variants = { path = "variants", version = "0.0.2" }
+variants-derive = { path = "derive", version = "0.0.2" }
+variants-macros = { path = "macros", version = "0.0.2" }
 # benchmarking
 criterion = { features = ["plotters"], version = "0.7" }
 # error handling
