+++ derive/src/attrs.rs
@@ -3,18 +3,42 @@
     authors: @FL03
 */
 #[doc(inline)]
-pub use self::prelude::*;
+pub use self::{nested::*, variant::*};
 
-pub mod nested;
-pub mod root;
-pub mod variants;
+mod nested;
+mod variant;
 
-pub(crate) mod prelude {
-    #[doc(inline)]
-    pub use super::nested::*;
-    #[doc(inline)]
-    #[allow(unused_imports)]
-    pub use super::root::*;
-    #[doc(inline)]
-    pub use super::variants::*;
+use syn::Attribute;
+
+/// The base interface for the `variant` attribute
+#[derive(Debug, Default)]
+pub struct OuterAttr {
+    pub variant: Option<VariantAttr>,
+}
+
+impl OuterAttr {
+    pub fn set_variant(&mut self, variant: VariantAttr) {
+        self.variant = Some(variant);
+    }
+
+    // tries to extract the scsys attribute from a list of attributes
+    pub fn extract(attrs: &[Attribute]) -> syn::Result<Self> {
+        let mut scsys = Self::default();
+        for attr in attrs {
+            if attr.path().is_ident("variants") {
+                attr.parse_nested_meta(|meta| {
+                    if let Ok(nested) = NestedAttr::parse_nested(&meta) {
+                        match nested {
+                            NestedAttr::Variant(inner) => {
+                                scsys.set_variant(inner);
+                                return Ok(());
+                            }
+                        }
+                    }
+                    Err(meta.error("unrecognized scsys attribute"))
+                })?;
+            }
+        }
+        Ok(scsys)
+    }
 }
